Add measuring option for distributed cpp/caf code

+++ measure_linux.py
@@ -19,12 +19,17 @@ def create_measurement_folder(folder_name, verbose):
         if verbose:
             print(f"[DEBUG] Folder '{folder_name}' already exists")
 
-def run_scaphandre(folder_name, file_name, regex, nano, verbose):
-    # Note: Make the regex more specific if possible.
+def run_scaphandre(folder_name, file_name, regex, nano, container, verbose):
     if nano > 0:
-        command = f"scaphandre json --step 0 --step-nano {nano} --process-regex={regex} -f {folder_name}/{file_name}.json"
+        if container:
+            command = f"scaphandre json --step 0 --step-nano {nano} --containers --container-regex={container} -f {folder_name}/{file_name}.json"
+        else:
+            command = f"scaphandre json --step 0 --step-nano {nano} --process-regex={regex} -f {folder_name}/{file_name}.json"
     else:
-        command = f"scaphandre json --step 1 --process-regex={regex} -f {folder_name}/{file_name}.json"
+        if container:
+            command = f"scaphandre json --step 1 --containers --container-regex={container} -f {folder_name}/{file_name}.json"
+        else:
+            command = f"scaphandre json --step 1 --process-regex={regex} -f {folder_name}/{file_name}.json"
     if verbose:
         print(f"[DEBUG] Running Scaphandre command: {command}")
     return subprocess.Popen(command, stdout=subprocess.PIPE if verbose else subprocess.DEVNULL,
@@ -40,21 +45,25 @@ def compile_program(program_exe, module):
         subprocess.run("cmake .", shell=True)
         subprocess.run("make", shell=True)
 
-def run_program(module, function, parameter, program_exe):
-    if program_exe == "beam.smp":
+def run_program(module, function, parameter, program_exe, container, index):
+    if program_exe == "beam.smp" and not container:
         erl_command = f"erl -noshell -run {module} {function} {parameter} -s init stop"
         print(f"[INFO] Running Erlang command: {erl_command}")
         return erl_command
     elif program_exe == "erlang":
-        distributed_erl_command = f"erl -noshell -sname 'runner@localhost' -setcookie cookie -run {module} {function} {parameter} -s init stop"
+        distributed_erl_command = f"erl -noshell -sname 'runner{index}@localhost' -setcookie cookie -run {module} {function} {parameter} -s init stop"
         print(f"[INFO] Running Distributed Erlang command: {distributed_erl_command}")
         return distributed_erl_command
+    elif container:
+        distributed_cpp_command = f"./{module}.sh {parameter}"
+        print(f"[INFO] Running Distributed C++ command: {distributed_cpp_command}")
+        return distributed_cpp_command
     else:
         cpp_command = f"./{module}.out {parameter}"
         print(f"[INFO] Running C++ command: {cpp_command}")
         return cpp_command
 
-def measure_energy_consumption(module, function, parameters, rep, nano, folder_name, result_file, program_exe, measure_cmd, verbose):
+def measure_energy_consumption(module, function, parameters, rep, nano, folder_name, result_file, program_exe, measure_cmd, container, verbose):
     for parameter in parameters:
         for i in range(rep):
             print(f"[INFO] Starting measurement for parameter '{parameter}' (Repetition {i+1}/{rep})")
@@ -65,20 +74,21 @@ def measure_energy_consumption(module, function, parameters, rep, nano, folder_n
 
             # Compile if necessary and prepare the execution command
             if not measure_cmd:
-                compile_program(program_exe, module)
-                command = run_program(module, function, parameter, program_exe)
+                if not container:
+                    compile_program(program_exe, module)
+                command = run_program(module, function, parameter, program_exe, container, i)
             else:
                 command = measure_cmd
                 print(f"[INFO] Using custom command: {command}")
 
             # Start Scaphandre
-            scaphandre_process = run_scaphandre(folder_name, file_name, r"\/" + program_exe, nano, verbose)
+            scaphandre_process = run_scaphandre(folder_name, file_name, r"\/" + program_exe, nano, container, verbose)
             time.sleep(5)
 
             # Run the command and measure the runtime
             print("[INFO] Starting program execution and timing...")
-            process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=True)
             start_time = timeit.default_timer()
+            process = subprocess.Popen(command, stdout=subprocess.PIPE, stderr=subprocess.STDOUT, shell=True)
             process.wait()
             end_time = timeit.default_timer()
             runtime = end_time - start_time
@@ -94,7 +104,7 @@ def measure_energy_consumption(module, function, parameters, rep, nano, folder_n
                 print("[WARNING] Scaphandre process did not terminate gracefully. Killing it...")
                 scaphandre_process.kill()
             subprocess.run('pkill -f scaphandre', shell=True)
-            if program_exe == "beam.smp":
+            if program_exe == "beam.smp" and not container:
                 subprocess.run(f'pkill -f {program_exe}', shell=True)
             time.sleep(5)
 
@@ -110,7 +120,7 @@ def measure_energy_consumption(module, function, parameters, rep, nano, folder_n
             number_samples = 0
 
             # Print the expected process name
-            if verbose:
+            if verbose and not container:
                 print(f"[DEBUG] Expected process name for energy consumption filtering: '{program_exe}'")
 
             for entry in data:
@@ -121,7 +131,7 @@ def measure_energy_consumption(module, function, parameters, rep, nano, folder_n
                     cmdline = consumer.get("cmdline", "")
                     consumption = consumer.get("consumption", 0.0)
 
-                    if (("scaphandre" not in exe and "dash" not in exe) and consumption > 0.0):
+                    if (("scaphandre" not in exe and "dash" not in exe)): # and consumption > 0.0):
                         if verbose:
                             print(f"[DEBUG] Matching process found: '{exe}', cmdline: '{cmdline}', adding consumption: {consumption}")
                         total_consumption += consumption
@@ -166,23 +176,25 @@ if __name__ == "__main__":
 
     # Define arguments
     parser.add_argument('module', type=str, help='module name (without extension for C++)')
-    parser.add_argument('function', type=str, help='function name (optional for C++)')
+    parser.add_argument('function', type=str, help='function name (can be anything for C++)')
     parser.add_argument('parameters', nargs='+', type=str, help='arguments for the function, separated by spaces')
     parser.add_argument('-c', '--cmd', type=str, default='', help='optional command to run the executable directly')
-    parser.add_argument('-e', '--exe', type=str, default='beam.smp', help='program process name (e.g., beam.smp for Erlang or the executable for C++)')
+    parser.add_argument('-e', '--exe', type=str, default='beam.smp', help='program process name (e.g., beam.smp or erlang for Erlang or the executable for C++)')
     parser.add_argument('-r', '--rep', type=int, default=1, help='number of repetitions')
     parser.add_argument('-f', '--file', type=str, default='', help='CSV result filename prefix')
     parser.add_argument('--folder', type=str, default='measurements', help='folder to store measurements')
+    parser.add_argument('--container', type=str, default='', help='containers to measure (for distributed C++ cases)')
     parser.add_argument('--nano', type=int, default=0, help='amount of nanoseconds between each measurement')
     parser.add_argument('-v', '--verbose', action='store_true', help='enable verbose output')
 
     # Parse arguments
     args = parser.parse_args()
 
+    # Create folder for measurements
     create_measurement_folder(args.folder, args.verbose)
 
     # Run the measurement with verbosity option
-    measure_energy_consumption(args.module, args.function, args.parameters, args.rep, args.nano, args.folder, args.file, args.exe, args.cmd, args.verbose)
+    measure_energy_consumption(args.module, args.function, args.parameters, args.rep, args.nano, args.folder, args.file, args.exe, args.cmd, args.container, args.verbose)
 
     # Exit the program
     sys.exit()
